src/containers/nmea_mux_demux/requirements.txt
nats-py==2.2
